Refactor (#1)
* Compile with Scala 2.12 (build is broken)

The code works with 2.11, and it builds with 2.12, but it is broken in 2.12. I need to investigate a little more on why it is broken. It looks like something to do with using asInstanceOf[T] to convert objects between JavaScript and Scala.

* Do not add 2.12 support yet

This is because 2.12 has bug around ThisFunction where passed "this" is being Scala-this instead of JS-this, which breaks it.